+++ constructors.go
@@ -4,7 +4,6 @@ import (
 	"net/http"
 
 	"github.com/alexedwards/scs/v2"
-	"github.com/trussworks/sesh/pkg/logger"
 )
 
 // NewUserSessions returns a configured UserSessions
@@ -12,7 +11,7 @@ func NewUserSessions(scs *scs.SessionManager, userDelegate UserDelegate, options
 
 	sessions := UserSessions{
 		scs,
-		logger.NewPrintLogger(),
+		newDefaultLogger(),
 		newDefaultErrorHandler(),
 		userDelegate,
 	}
@@ -27,8 +26,12 @@ func NewUserSessions(scs *scs.SessionManager, userDelegate UserDelegate, options
 	return sessions, nil
 }
 
+// Option is an option for constructing a UserSessionManager, they can be passed in to NewUserSessions
+// The available options are defined below.
 type Option func(*UserSessions) error
 
+// CustomLogger supplies a custom logger for logging session lifecycle events.
+// It must conform to EventLogger
 func CustomLogger(logger EventLogger) Option {
 	return func(userSessions *UserSessions) error {
 		userSessions.logger = logger
@@ -36,6 +39,8 @@ func CustomLogger(logger EventLogger) Option {
 	}
 }
 
+// CustomErrorHandler supplies a custom http.Handler for responding to errors in the ProtectedMiddleware
+// Use ErrorFromContext(ctx) to get the error that caused this handler to be called.
 func CustomErrorHandler(errorHandler http.Handler) Option {
 	return func(userSessions *UserSessions) error {
 		userSessions.errorHandler = errorHandler
